A / B

+++ baekjoon/1008.c
@@ -5,6 +5,6 @@ int main()
     int A, B;
 
     scanf("%d %d", &A, &B);
-    printf("%f", (float) A / B);
+    printf("%.9f", (double) A / B);
     return 0;
 }
